update and improve the codes

+++ utils.py
@@ -1,19 +1,19 @@
 import os
-os.environ["CUDA_VISIBLE_DEVICES"]="-1"
-import tensorflow as tf
-from tensorflow.python import graph_util
-import numpy as np
+import warnings
 from shutil import rmtree
+
+import numpy as np
+import tensorflow as tf
 
 
 def validate(path):
-    path = os.path.abspath(path) + '/'
+    path = os.path.abspath(path)
     if not os.path.exists(path):
         os.mkdir(path)
     return path
 
 
-def find_all_files(path, extensions=[], exclude=[]):
+def find_all_files(path, extensions=(), exclude=()):
     # Find all images with given extensions
     path = os.path.abspath(path) + '/'
     if len(extensions) > 0:
@@ -26,22 +26,12 @@ def find_all_files(path, extensions=[], exclude=[]):
     return files
 
 
-def load_graph(graph_path, return_elements=[]):
+def load_graph(graph_path, return_elements=None):
     # Creates graph from saved graph_def.pb
-    with tf.gfile.GFile(graph_path, 'rb') as f:
-        graph_def = tf.GraphDef()
+    with tf.io.gfile.GFile(graph_path, 'rb') as f:
+        graph_def = tf.compat.v1.GraphDef()
         graph_def.ParseFromString(f.read())
-        output_nodes = tf.import_graph_def(graph_def, return_elements=return_elements)
-        return output_nodes
-
-
-def load_graph2(graph_path, inp, inp_name='/data', out_name='/prob:0', graph_name='model'):
-    # Creates graph from saved graph_def.pb.
-    with tf.gfile.GFile(graph_path, 'rb') as f:
-        graph_def = tf.GraphDef()
-        graph_def.ParseFromString(f.read())
-        tf.import_graph_def(graph_def, input_map={inp_name: inp}, name=graph_name)
-        output_nodes = tf.get_default_session().graph.get_tensor_by_name(graph_name + out_name)
+        output_nodes = tf.compat.v1.import_graph_def(graph_def, return_elements=return_elements)
         return output_nodes
 
 
@@ -49,38 +39,23 @@ def freeze_save_graph(sess, name, output_node, log_dir):
     for node in sess.graph.as_graph_def().node:
         node.device = ""
     variable_graph_def = sess.graph.as_graph_def()
-    optimized_net = graph_util.convert_variables_to_constants(sess, variable_graph_def, [output_node])
-    tf.train.write_graph(optimized_net, log_dir, name, False)
+    optimized_net = tf.compat.v1.graph_util.convert_variables_to_constants(sess, variable_graph_def, [output_node])
+    tf.io.write_graph(optimized_net, log_dir, name, False)
 
 
-def next_batch(data, batch_size):
+def next_batch(data, batch_size, shuffle=False):
+    """
+    :param data: list or array
+    :param batch_size: int, the size of the batch
+    :param shuffle: bool, shuffle data before selecting the batch
+    :return: tuple, (remaining data, batch data)
+    """
     if len(data) <= batch_size:
         return [], data
     else:
+        if shuffle:
+            np.random.shuffle(data)
         return data[batch_size:], data[:batch_size]
-
-
-def next_batch_shuffle(data, batch_size):
-    """
-    :param data: 1D array
-    :return:
-    """
-    if len(data) <= batch_size:
-        return [], np.random.choice(data, batch_size)
-    else:
-        np.random.shuffle(data)
-        return data[batch_size:], data[:batch_size]
-
-
-def next_batch_shuffle_nd(data, batch_size):
-    if len(data) <= batch_size:
-        np.random.shuffle(data)
-        return [], data
-    else:
-        mask = np.zeros(len(data), bool)
-        inds = np.random.choice(range(len(mask)), batch_size, False)
-        mask[inds] = True
-        return data[mask], data[~mask]
 
 
 def save(sess, saver, model_path, model_name):
@@ -102,8 +77,8 @@ def load(sess, saver, model_path):
 
 def clear_start(paths):
     for path in paths:
-        validate(path)
-        rmtree(path)
+        if os.path.isdir(path):
+            rmtree(path)
         validate(path)
 
 
@@ -113,16 +88,14 @@ def variable_summaries(var, name):
         std = tf.sqrt(tf.reduce_mean(tf.square(var - mean)))
         mx = tf.reduce_max(var)
         mn = tf.reduce_min(var)
-    tf.summary.scalar(name + '/mean', mean)
-    tf.summary.scalar(name + '/stddev', std)
-    tf.summary.scalar(name + '/max', mx)
-    tf.summary.scalar(name + '/min', mn)
-    tf.summary.histogram(name, var)
+    tf.compat.v1.summary.scalar(name + '/mean', mean)
+    tf.compat.v1.summary.scalar(name + '/stddev', std)
+    tf.compat.v1.summary.scalar(name + '/max', mx)
+    tf.compat.v1.summary.scalar(name + '/min', mn)
+    tf.compat.v1.summary.histogram(name, var)
 
 
-def construct_label(i, num_classes):
-    l = np.zeros(num_classes)
-    l[i] = 1
-    return l
-
-
+def construct_label(index, num_classes):
+    label = np.zeros(num_classes)
+    label[index] = 1
+    return label
